+++ apps/oxfmt/src-js/cli-worker.ts
@@ -1,0 +1,4 @@
+// `oxfmt` CLI - Worker Thread Entry Point
+
+// Re-exports core functions for use in `worker_threads`
+export { formatEmbeddedCode, formatFile } from "./libs/prettier";
