+++ apps/oxfmt/src-js/cli/worker-proxy.ts
@@ -1,0 +1,33 @@
+import Tinypool from "tinypool";
+import { resolvePlugins } from "../libs/prettier";
+import type { Options } from "prettier";
+
+// Worker pool for parallel Prettier formatting
+let pool: Tinypool | null = null;
+
+export async function initExternalFormatter(numThreads: number): Promise<string[]> {
+  pool = new Tinypool({
+    filename: new URL("./cli-worker.js", import.meta.url).href,
+    minThreads: numThreads,
+    maxThreads: numThreads,
+  });
+
+  return resolvePlugins();
+}
+
+export async function formatEmbeddedCode(
+  options: Options,
+  tagName: string,
+  code: string,
+): Promise<string> {
+  return pool!.run({ options, code, tagName }, { name: "formatEmbeddedCode" });
+}
+
+export async function formatFile(
+  options: Options,
+  parserName: string,
+  fileName: string,
+  code: string,
+): Promise<string> {
+  return pool!.run({ options, code, fileName, parserName }, { name: "formatFile" });
+}
