+++ apps/oxfmt/src-js/libs/prettier.ts
@@ -1,0 +1,96 @@
+import type { Options } from "prettier";
+
+// Lazy load Prettier
+//
+// NOTE: In the past, statically importing caused issues with `oxfmt --lsp` not starting.
+// However, this issue has not been observed recently, possibly due to changes in the bundling configuration.
+// Anyway, we keep lazy loading for now to minimize initial load time.
+let prettierCache: typeof import("prettier");
+
+/**
+ * TODO: Plugins support
+ * - Read `plugins` field
+ * - Load plugins dynamically and parse `languages` field
+ * - Map file extensions and filenames to Prettier parsers
+ *
+ * @returns Array of loaded plugin's `languages` info
+ */
+export async function resolvePlugins(): Promise<string[]> {
+  return [];
+}
+
+const TAG_TO_PARSER: Record<string, string> = {
+  // CSS
+  css: "css",
+  styled: "css",
+  // GraphQL
+  gql: "graphql",
+  graphql: "graphql",
+  // HTML
+  html: "html",
+  // Markdown
+  md: "markdown",
+  markdown: "markdown",
+};
+
+/**
+ * Format xxx-in-js code snippets
+ *
+ * @returns Formatted code snippet
+ * TODO: In the future, this should return `Doc` instead of string,
+ * otherwise, we cannot calculate `printWidth` correctly.
+ */
+export async function formatEmbeddedCode({
+  code,
+  tagName,
+  options,
+}: {
+  code: string;
+  tagName: string;
+  options: Options;
+}): Promise<string> {
+  // TODO: This should be resolved in Rust side
+  const parserName = TAG_TO_PARSER[tagName];
+
+  // Unknown tag, return original code
+  if (!parserName) return code;
+
+  if (!prettierCache) {
+    prettierCache = await import("prettier");
+  }
+
+  // SAFETY: `options` is created in Rust side, so it's safe to mutate here
+  options.parser = parserName;
+  return prettierCache
+    .format(code, options)
+    .then((formatted) => formatted.trimEnd())
+    .catch(() => code);
+}
+
+/**
+ * Format non-js file
+ *
+ * @returns Formatted code
+ */
+export async function formatFile({
+  code,
+  parserName,
+  fileName,
+  options,
+}: {
+  code: string;
+  parserName: string;
+  fileName: string;
+  options: Options;
+}): Promise<string> {
+  if (!prettierCache) {
+    prettierCache = await import("prettier");
+  }
+
+  // SAFETY: `options` is created in Rust side, so it's safe to mutate here
+  // We specify `parser` to skip parser inference for performance
+  options.parser = parserName;
+  // But some plugins rely on `filepath`, so we set it too
+  options.filepath = fileName;
+  return prettierCache.format(code, options);
+}
