--- apps/oxfmt/src-js/migration/index.ts
@@ -1,2 +1,0 @@
-export * from "./init.js";
-export * from "./migrate-prettier.js";
